+++ sbk-api/src/main/java/io/sbk/perl/ReportLatency.java
@@ -14,11 +14,10 @@ public interface ReportLatency {
 
     /**
      * Record the latency.
-     *
-     * @param startTime start time.
+     *  @param startTime start time.
      * @param bytes number of bytes.
      * @param events number of events(records).
      * @param latency latency value in milliseconds.
      */
-    void recordLatency(long startTime, long bytes, long events, long latency);
+    void recordLatency(long startTime, int bytes, int events, long latency);
 }
