+++ sbk-api/src/main/java/io/sbk/perl/impl/HashMapLatencyRecorder.java
@@ -106,14 +106,13 @@ public class HashMapLatencyRecorder extends LatencyRecordWindow {
 
     /**
      * Record the latency.
-     *
-     * @param startTime start time.
+     *  @param startTime start time.
      * @param bytes number of bytes.
      * @param events number of events(records).
      * @param latency latency value in milliseconds.
      */
     @Override
-    public void recordLatency(long startTime, long bytes, long events, long latency) {
+    public void recordLatency(long startTime, int bytes, int events, long latency) {
         if (record(bytes, events, latency)) {
             reportLatency(latency, events);
         }
